pyproject.toml
[project]
name = "NetworkSecuritySystemMLProject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "certifi>=2026.1.4",
    "dagshub>=0.6.4",
    "dill>=0.4.0",
    "fastapi>=0.128.0",
    "mlflow>=3.8.1",
    "numpy>=2.4.1",
    "pandas>=2.3.3",
    "pyaml>=25.7.0",
    "pymongo[srv]==3.6",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
]
